created CoreseShacl wrapper
